Implement read_solution
