+++ cvrplib/__init__.py
@@ -1,2 +1,2 @@
 from .download import download_instance, download_solution, list_names
-from .read import read, read_instance, read_solution
+from .read import read_instance, read_solution
