+++ cvrplib/read/__init__.py
@@ -1,2 +1,2 @@
-from .read import read
 from .read_instance import read_instance
+from .read_solution import read_solution
